try to fix session expred notification
